Add world flag + fix get stats service
Add wolrd flag to show it for all players without a country flag

Fix stats service:

When the array fo matches to calculate stats from is 0, the reduce()
method caused errors. Validation added to return 0 instead of calculate.

The same happened when calculating the percentge of victories when the
array of matches was empty.